feat: moved Buy Me Coffee content into his own page
