+++ apps/build-onchain-apps/src/pages/_document.tsx
@@ -3,11 +3,7 @@ import { Html, Head, Main, NextScript } from 'next/document';
 export default function DocumentPage() {
   return (
     <Html lang="en">
-      <Head>
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-        <link rel="manifest" href="/manifest.json" />
-      </Head>
+      <Head />
       <body>
         <Main />
         <NextScript />
